code review suggestion
