+++ src/components/JsonView.vue
@@ -1,13 +1,13 @@
 <template>
-  <code class="json-view" :class="classes">
-    <!-- eslint-disable-next-line vue/no-v-html -->
-    <pre class="json-view__lines"><template v-for="(line, index) in lines" :key="index"><span class="json-view__line" v-html="line" /></template></pre>
+  <code class="json-view">
+    <CodeLines :code="code" :show-line-numbers="showLineNumbers" />
   </code>
 </template>
 
 <script lang="ts" setup>
   import { highlight, languages } from 'prismjs'
   import { computed } from 'vue'
+  import CodeLines from '@/components/CodeLines.vue'
   import { definition } from '@/utilities/languageDefinitions/json'
 
   languages.json = definition
@@ -17,13 +17,9 @@
     showLineNumbers?: boolean,
   }>()
 
-  const lines = computed(() => {
-    return highlight(props.value ?? '', languages.json, 'json').split('\n')
+  const code = computed(() => {
+    return highlight(props.value ?? '', languages.json, 'json')
   })
-
-  const classes = computed(() => ({
-    'json-view--with-line-numbers': props.showLineNumbers,
-  }))
 </script>
 
 <style>
@@ -43,28 +39,6 @@
   break-normal
   rounded
   text-left;
-}
-
-.json-view__lines { @apply
-  py-1;
-  counter-reset: line;
-}
-
-.json-view__line { @apply
-  h-5
-  w-full
-  block
-}
-
-.json-view--with-line-numbers .json-view__line:before { @apply
-  inline-block
-  px-3
-  mr-5
-  border-r-[1px]
-  border-slate-200
-  text-slate-400;
-  counter-increment: line;
-  content: counter(line);
 }
 
 .json-view .token.comment,
